feat!: Adding gateway service & client access policy

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -1,6 +1,10 @@
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
+using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -8,13 +12,19 @@ namespace AuctionService.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class AuctionsController(AuctionDbContext context, IMapper mapper) : ControllerBase
+public class AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint) : ControllerBase
 {
     [HttpGet]
-    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions()
+    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string? date)
     {
-        var auctions = await context.Auctions.Include(x => x.Item).OrderBy(x => x.Item.Make).ToListAsync();
-        return mapper.Map<List<AuctionDto>>(auctions);
+        var query = context.Auctions.OrderBy(x => x.Item.Make).AsQueryable();
+
+        if (!string.IsNullOrEmpty(date))
+        {
+            query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
+        }
+
+        return await query.ProjectTo<AuctionDto>(mapper.ConfigurationProvider).ToListAsync();
     }
 
     [HttpGet("{id}")]
@@ -25,44 +35,56 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper) : Cont
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = mapper.Map<Auction>(createAuctionDto);
 
-        auction.Seller = "test";
+        auction.Seller = User.Identity?.Name!;
+
         context.Auctions.Add(auction);
 
+        var newAuction = mapper.Map<AuctionDto>(auction);
+        await publishEndpoint.Publish(mapper.Map<AuctionCreated>(newAuction));
         var result = await context.SaveChangesAsync() > 0;
+
         if (!result) return BadRequest("Couldn't save changes to the DB");
 
-        return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, mapper.Map<AuctionDto>(auction));
+        return CreatedAtAction(nameof(GetAuctionById), new { newAuction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
-        Console.WriteLine("Hello world");
         var auction = await context.Auctions.Include(a => a.Item).FirstOrDefaultAsync(a => a.Id == id);
         if (auction == null) return NotFound();
+
+        Console.WriteLine(User.Identity?.Name);
+        Console.WriteLine("Forbidden Request");
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
         auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
         auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
-
+        await publishEndpoint.Publish(mapper.Map<AuctionUpdated>(auction));
         var result = await context.SaveChangesAsync() > 0;
         if (result) return Ok();
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await context.Auctions.FindAsync(id);
         if (auction == null) return NotFound();
+        if (auction.Seller != User.Identity?.Name) return Forbid();
         context.Auctions.Remove(auction);
+        await publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
         var result = await context.SaveChangesAsync() > 0;
         if (!result) BadRequest("Couldn't not update DB");
         return Ok();
